solved TS compile issue, tested table and metric routes - both work successfully
