+++ server/package.json
@@ -1,10 +1,11 @@
 {
   "name": "server",
   "version": "1.0.0",
-  "main": "server.js",
+  "main": "server.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "nodemon server.js"
+    "start": "nodemon --exec ts-node server.ts",
+    "start-test": "ts-node-dev --respawn server.ts"
   },
   "author": "",
   "license": "ISC",
@@ -13,17 +14,17 @@
     "@aws-sdk/client-bedrock-runtime": "^3.613.0",
     "@aws-sdk/client-cloudwatch": "^3.609.0",
     "@aws-sdk/client-dynamodb": "^3.609.0",
-    "@types/express": "^4.17.21",
-    "@types/node": "^20.14.10",
     "bcrypt": "^5.1.1",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
-    "react-cookie": "^7.1.4",
-    "ts-node": "^10.9.2",
-    "typescript": "^5.5.3"
+    "react-cookie": "^7.1.4"
   },
   "devDependencies": {
-    "nodemon": "^3.1.4"
-  },
-  "type": "module"
+    "@types/express": "^4.17.21",
+    "@types/node": "^20.14.10",
+    "nodemon": "^3.1.4",
+    "ts-node": "^10.9.2",
+    "ts-node-dev": "^2.0.0",
+    "typescript": "^5.5.3"
+  }
 }
